requirements.txt
fastapi==0.115.6
sqlmodel==0.0.22
pydantic==2.10.4
pymysql==1.1.1
uvicorn==0.34.0
pyjwt==2.10.1
passlib[bcrypt]==1.7.4
